Feat: Fixed CORS issue and deploying to vercel
