+++ app/layout.js
@@ -1,3 +1,4 @@
+// "use client";
 import { Inter, Montserrat } from "next/font/google";
 import "./globals.css";
