+++ app/page.js
@@ -4,14 +4,18 @@ import FooterNav from "@/components/common/FooterNav";
 import Header from "@/components/common/Header";
 import MainContainer from "@/components/common/MainContainer";
 import { MetamaskProvider } from "../hooks/useMetamask";
+import { AppProps } from "next/app";
 
-export default function Home() {
+export default function Home({ Component, pageProps }) {
   return (
-    <MainContainer marginTop={"main"}>
-      <Header title={"hi casmir ✌🏽"} />
-      <HomeHero />
-      <HomeMain />
-      <FooterNav />
-    </MainContainer>
+    <MetamaskProvider>
+      <MainContainer marginTop={"main"}>
+        <Header title={"hi casmir ✌🏽"} />
+        <HomeHero />
+        <HomeMain />
+        <FooterNav />
+      </MainContainer>
+      <Component {...pageProps} />
+    </MetamaskProvider>
   );
 }
