+++ server.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const fetch = require("node-fetch");
 const cors = require("cors"); // Import the cors package
-const DataFetching = require("./components/DataFetching"); // Assuming your DataFetching component is in the same directory
+const DataFetching = require("./components/DataFetcher"); // Assuming your DataFetching component is in the same directory
 
 const app = express();
 const port = 3000;
